Refactor testing tools: move to separate package, split file contents into dedicated files.
